web-app/backend/requirements.txt
uvicorn
fastapi
numpy
opencv-python
torch
torchvision
Pillow
segmentation-models-pytorch
tensorflow
